+++ cmd/nofox/main.go
@@ -2,17 +2,30 @@ package main
 
 import (
 	"flag"
+	"fmt"
 	"io"
 	"log"
 	"os"
+	"runtime/debug"
 
 	"github.com/cpmachado/nofox"
 )
 
 func main() {
 	filename := ""
+	version := false
 	flag.StringVar(&filename, "f", "", "bf script to load")
+	flag.BoolVar(&version, "v", version, "display version")
 	flag.Parse()
+
+	if version {
+		vsn := "dev"
+		if info, ok := debug.ReadBuildInfo(); ok {
+			vsn = info.Main.Version
+		}
+		fmt.Printf("nofox-%s\n", vsn)
+		os.Exit(0)
+	}
 
 	if filename == "" {
 		flag.Usage()